+++ tests/testthat/test-data-storage-postgresql.R
@@ -1,3 +1,3 @@
 
 # Test suite common to data storages (see `helper-data_storage.R`)
-test_that_common_data_storage(init_test_sqlite, "SQLite")
+test_that_common_data_storage(init_test_postgres, "PostgreSQL")