+++ NEWS.md
@@ -1,8 +1,10 @@
 # shiny.telemetry (development version)
 
-- Fix the way of getting the session token.
-- Allow optional username overwrite.
+- Fixed the way of getting the session token.
+- Allowed optional username overwrite.
 - Added MS SQL Server support (see `DataStorageMSSQLServer` class).
+- Added CI tests all DBI-based DataStorage providers.
+- Added optional parameter to `read_event_data` that filters by `app_name`.
 
 # shiny.telemetry 0.1.0